+++ irc/logging.go
@@ -1,0 +1,14 @@
+package irc
+
+import (
+	"github.com/inconshreveable/log15"
+)
+
+var (
+	Logger log15.Logger
+)
+
+func init() {
+	Logger = log15.New()
+	Logger.SetHandler(log15.StdoutHandler)
+}