+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# 1.0 (not released)
+# 0.10.0
 
-* Added floor protection for connections. Just add "floodProtection": true to
-  connection config.
+* Added backoff to failed connections when they try to reconnect. This replaces
+  the retries setting that will stop retrying when the max is reached.
+* Kind of fixed a bug in the connection watchdog that would try to send a PING
+  on a nil channel when a connection was disconnected.
+* Fixed logging. It was terrible. It's still not great, but it's much better.